python-backend/pyproject.toml
[project]
name = ""
version = "0.0.1"
dependencies = [
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "pylint>=3.3.7",
    "python-multipart>=0.0.6",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.1.0",
    "psycopg>=2.9.9",
    "python-dotenv>=1.0.0",
    "langchain-core>=0.3.76",
    "sqlmodel>=0.0.23",
    "alembic>=1.16.4",
    "passlib==1.7.4",
    "pyjwt>=2.10.0",
    "bcrypt~=3.2.0",
    "resend>=2.13.0",
    "httpx>=0.28.0",
    "pystache>=0.6.7",
    "redis>=6.3.0",
    "langchain-google-genai>=2.1.10",
    "langchain-groq>=0.3.7",
    "langchain>=0.3.26",
    "langchain-core==0.3.76",
    "langchain-tavily>=0.2.10",
    "psycopg2-binary>=2.9.10",
    "duckduckgo-search>=8.1.1",
    "bs4>=0.0.2",
    "langchainhub>=0.1.21",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]
[dependency-groups]
dev = ["Workers-py"]

[tool.uv]
package = false
